dom/tests/mochitest/webcomponents/test_document_register.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=783129
-->
<head>
  <title>Test for document.register using custom prototype</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script>
var gElementUpgraded = false;
var gElementReplaced = false;

function elementUpgrade() {
  gElementUpgraded = true;

  // Check for prototype on upgraded element.
  var documentElement = document.getElementById("grabme");
  ok(documentElement.hello, "Upgraded element should inherit 'hello' method from prototype.");
  documentElement.hello();

  var customChild = document.getElementById("kid");
  ok(customChild, "Upgrade should preserve children.");
  ok(customChild.parentNode == documentElement, "Parent should be updated to new custom element.");

  // Try creating new element and checking for prototype.
  var constructedElement = document.createElement("x-hello");
  ok(constructedElement.hello, "Created element should inherit 'hello' method from prototype.");
  constructedElement.hello();
}

function elementReplace(e) {
  gElementReplaced = true;

  ok(e.upgrade != e.target, "Upgraded element should be different from the target.");
  ok(e.upgrade.firstElementChild.id == "kid", "Upgrade element should have a child.");
  ok(e.target.firstElementChild.id == "kid", "Replacement element should have a child.");
}

function startTest() {
  // Create a prototype that inheits from HTMLElement.
  var customProto = {};
  customProto.__proto__ = HTMLElement.prototype;
  customProto.hello = function() {
    ok(true, "Custom element should use provided prototype.");
  };

  var oldElem = document.getElementById("grabme");
  oldElem.addEventListener("elementreplace", elementReplace);

  document.addEventListener("elementupgrade", elementUpgrade);
  var elementConstructor = document.register("x-hello", { prototype: customProto });

  // Try creating new element and checking for prototype.
  var constructedElement = new elementConstructor();
  ok(constructedElement.hello, "Created element should inherit 'hello' method from prototype.");
  constructedElement.hello();

  ok(!oldElem.hello, "Element obtained prior to registration should not have inherited prototype.");

  ok(gElementUpgraded && gElementReplaced, "Upgrade and replace events should have been fired.");

  // Check that custom elements registered without a prototype inherit from HTMLElement.
  document.register("x-bye");
  var byeElement = document.createElement("x-bye");
  ok(byeElement.__proto__.__proto__, HTMLElement.prototype, "Element registered without prototype should inherit from HTMLElement.");

  // Check that element registration with a prototype that does not inherit from HTMLElement results in exception.
  try {
    document.register("x-foo", { "prototype": {} });
    ok(false, "Prototype that does not inherit from HTMLElement should throw exception");
  } catch (ex) {
    ok(true, "Prototype that does not inherit from HTMLElement should throw exception");
  }

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
</script>

</head>
<body onload="startTest()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=783129">Bug 783129</a>
<x-hello id="grabme">
<div id="kid"></div>
</x-hello>
</body>
</html>
